Isolate all configuration-related parameters in a configuration sub-module to highlight the actually relevant types on the homepage of the documentation for the crate.
